+++ src/utils.ts
@@ -3,6 +3,7 @@ import * as url from "url";
 import * as http from "http";
 import * as clc from "cli-color";
 import * as ora from "ora";
+import * as process from "process";
 import { Readable } from "stream";
 import * as winston from "winston";
 import { SPLAT } from "triple-beam";
@@ -446,4 +447,34 @@ export function createDestroyer(server: http.Server): () => Promise<void> {
     }
     return destroyPromise;
   };
+}
+
+/**
+ * Returns the given date formatted as `YYYY-mm-dd HH:mm:ss`.
+ * @param d the date to format.
+ * @return the formatted date.
+ */
+export function datetimeString(d: Date): string {
+  const day = `${d.getFullYear()}-${(d.getMonth() + 1).toString().padStart(2, "0")}-${d
+    .getDate()
+    .toString()
+    .padStart(2, "0")}`;
+  const time = `${d
+    .getHours()
+    .toString()
+    .padStart(2, "0")}:${d
+    .getMinutes()
+    .toString()
+    .padStart(2, "0")}:${d
+    .getSeconds()
+    .toString()
+    .padStart(2, "0")}`;
+  return `${day} ${time}`;
+}
+
+/**
+ * Indicates whether the end-user is running the CLI from a cloud-based environment.
+ */
+export function isCloudEnvironment() {
+  return !!process.env.CODESPACES;
 }