+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- Change default runtime for Firebase Extensions template to nodejs10.
-- Modify messages for Firebase Extensions Node.js 10 migration.
+- Replaces `tslint` with `eslint` as the default linter in new Cloud Functions for Firebase directories. (#1663)